requirements.txt
patch
svn
git-python==1.0.3